% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TOA_simulations_1region}
\alias{TOA_simulations_1region}
\title{TOA MAIN DATASET: Antarctic Toothfish 1 Region Simulations}
\format{
A list object.
}
\usage{
data(TOA_simulations_1region)
}
\description{
1 region simulations of TOA from SPM OM.
}
\examples{
data(TOA_simulations_1region)
head(realised.catches)
head(realised.tags)
}
\keyword{dataset}
