% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TOA_simulations_cells}
\alias{TOA_simulations_cells}
\title{TOA SUPPLEMENTAL DATASET: Antarctic Toothfish Spatial Simulations}
\format{
A list object.
}
\usage{
data(TOA_simulations_cells)
}
\description{
Spatial simulations of TOA from SPM OM.
}
\examples{
data(TOA_simulations_cells)
head(sim)
}
\keyword{dataset}
