% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TOA_simulations_inputs_4regions}
\alias{TOA_simulations_inputs_4regions}
\title{TOA MAIN DATASET: Antarctic Toothfish 4 Region Simulations}
\format{
A list object.
}
\usage{
data(TOA_simulations_inputs_4regions)
}
\description{
4 region simulations of TOA from SPM OM.
}
\examples{
data(TOA_simulations_inputs_4regions)
head(realised.catches)
head(realised.tags)
}
\keyword{dataset}
