% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YFT_1area_observations_1_100}
\alias{YFT_1area_observations_1_100}
\title{100 simulated YFT datasets at the 1 area scale}
\format{
A data frame 'YFT_1area_observations_1_100.RData' with variables:
\describe{
\item{dat_1A_X}{Simulated aggregated 1 area datasets for YFT; X=1:100}
\item{biol_dat}{Biological data for YFT}
}
}
\usage{
data('YFT_1area_observations_1-100')
}
\description{
A dataset 100 simulations of the YFT 1 area aggregated catch, cpue, length frequencies, and tagging data (along with biological data).
}
\details{
each sim list with following:
}
\examples{
data('YFT_1area_observations_1-100')
head(biol_dat)

}
