% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YFT_221cell_observations_1_33}
\alias{YFT_221cell_observations_1_33}
\title{SUPPLEMENTAL DATASET: Fully spatial simulated datasets
(Simulation #: 1-33)}
\format{
A data frame 'YFT_221cell_observations_1_33' with variables:
\describe{
\item{sim_1}{Simulated aggregated 1 area datasets for YFT; Simulation #1-33}
\item{biol_dat}{Biological data for YFT}
}
}
\usage{
data('YFT_221cell_observations_1_33')
}
\description{
This R data file contains 1-33 simulated data sets at the 221 5x5 gird cell level, which
matches the spatial resolution used in the OM.  These data are provided for your
information and in a usable format if participants wish to develop spatially
structured models beyond the one and four area aggregations already provided.
}
\section{The key lists in \strong{sim_X$} (e.g., sim_1$) include:}{
\emph{\strong{Sim_X}} (X: sim # 1-100)\preformatted{     ***obs***: Simulated observations
     
           ***simulated_ll_jpn_cpue_YYY***:
           
           ***simulated_XXX_lf_YYY***:
           
     ***tagrel***: Tag releases
     
           ***process[tag_XXX]***: tag release in a given pseudo-year
           
     ***tagrecs***: Simulated tag recapture data
     
           ***observation[tag_recapture_XXX_in_YYY]***: tag recapture in a given pseudo-year from release cohort XXX
           
     ***layer***
     
           ***base***: 1 = on the water, 0 = on land
           
           ***cell***: unique cell row and column numbers
           
           ***constant***: 1
           
           ***latitude***: latitude for each cell center
           
           ***longitude***: longitude for each cell center
           
           ***region***: IOTC YFT regions (R1a, R1b, R2, R3, R4, R5)
           
           ***sst***: Sea Surface Temperature (see Dunn et al. 2020)
           
           ***clo***: Chlorophyll (see Dunn et al. 2020)
           
           ***fishing_ff_YY***: ff: fishery = ps, trol, bb, gill, ll, other, hand, YY = pseudo-year
}
}

\section{The lists in \strong{biol_dat} include:}{
\emph{M:} age varying natural mortality

\emph{age}: first age to last age (pseudo-years)

\emph{maturity}: maturity at age ogive (pseudo-years)

\emph{k}: age varying growth coefficients

\emph{L}: length (cm)

\emph{Linf}: length infinity (cm)

\emph{Lmin}: length minimum (cm)

\emph{a}: length-weight scaling coefficient

\emph{b}: length-weight shape parameter
}

\examples{
data('YFT_221cell_observations_1_33')
head(sim_1$obs$simulated_cpue_ll_jpn_86$data$obs)
}
