% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YFT_4area_observations_1-100}
\alias{YFT_4area_observations_1-100}
\title{YFT_4area_observations_1-100}
\format{
A data frame 'YFT_4area_observations_1-100.RData' with variables:
\describe{
\item{dat_4A_X}{Simulated aggregated 4 area datasets for YFT; X=1:100}
\item{biol_dat}{Biological data for YFT}
}
}
\usage{
data('YFT_4area_observations_1-100')
}
\description{
A dataset 100 simulations of the YFT 1 area aggregated catch, cpue, length frequencies, and tagging data (along with biological data).
}
\details{
each sim list with following:
}
\examples{
data('YFT_4area_observations_1-100')
head(biol_dat)
}
