% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YFT_SRD_1A_4}
\alias{YFT_SRD_1A_4}
\title{YFT_4area_observations_1-100}
\format{
A data frame 'YFT_4area_observations_1-100.RData' with variables:
\describe{
\item{YFT_SRD_1A_4}{Simulated aggregated 1 area datasets for YFT; Simulation #4}
\item{biol_dat}{Biological data for YFT}
}
}
\usage{
data('YFT_SRD_1A_4')
}
\description{
A dataset 100 simulations of the YFT 1 area aggregated catch, cpue, length frequencies, and tagging data (along with biological data).
}
\details{
each sim list with following:
}
\examples{
data('YFT_SRD_1A_4')
head(biol_dat)
}
