% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YFT_SRD_4A_4_v2}
\alias{YFT_SRD_4A_4_v2}
\title{START HERE (YFT): 4 area single representative dataset}
\format{
A data frame 'YFT_SRD_4A_4_v2.RData' with variables:
\describe{
\item{dat_4A_X}{Simulated aggregated 4 area datasets for YFT}
\item{biol_dat}{Biological data for YFT}
}
}
\usage{
data('YFT_SRD_4A_4_v2')
}
\description{
A single representative dataset aggregated at the 4 area scale. The chosen dataset is simulation #4.
List objects include aggregated catch, cpue, length frequencies, and tagging data (along with biological data).
}
\section{The key lists in \strong{YFT_SRD_4A_4_v2$} (e.g., dat_4A_4$lencomp) include:}{
\emph{\strong{lencomp}}: (list) 4 area dataframe of aggregated length frequencies by length bin with columns corresponding to pseudo-year (Yr), season(Seas), fleet/survey index (FltSvy), sex (Gender),  partition – length data come from all fish encountered (Part=0), number of lengths sampled (Nsamp), and the set of length bins.

\emph{\strong{catch}}: (list) 4 area dataframe of catch by fishery, including pseudo-year (year), and    season (seas)

\emph{\strong{CPUE/cpu}}: (list) 4 area cpue with columns for pseudo-year (year), season (seas), cpue (cpu), standard deviation associated with lognormal error in cpue (cv), and survey index (index); note that CPUE is for longline only

\emph{\strong{tag_releases}}: 4 area tag release data with columns corresponding to generic tag release event number (tg), area of release (reg), pseudo-year (yr), season (season), sex (Gender), age (age), and number of releases (nrel) adjusted for initial tag mortality

\emph{\strong{tag_recaps}}: 4 area tag recapture data with columns corresponding to generic tag release event number (tg), pseudo-year (yr), season (season), fleet index (fleet), and the number of recaptures (recaps)
}

\section{Explanations of list object items in the provided data files that analysts may find additionally useful.}{
Much of these follow, or closely follow, Stock Synthesis (version SS3.24Z) naming conventions.  The name is the list identifier in each Rdata object (e.g., the name styr is found using dat_1A_4$styr).

\emph{\strong{styr}}	Start year (pseudo-year)

\emph{\strong{endyr}}	End year (pseudo-year)

\emph{\strong{nseas}}	Number of seasons

\emph{\strong{spawn_seas}}	Season when spawning occurs

\emph{\strong{Nfleet}}	Number of fishing fleets

\emph{\strong{Nsurveys}}	Number of surveys (fishery-dependent CPUE)

\emph{\strong{N_areas}}	Number of spatial areas

\emph{\strong{fleetnames}}	Specific fleet and survey index names

\emph{\strong{surveytiming}}	Timing of fleet and surveys

\emph{\strong{areas}}	Areas where each fleet or survey operate

\emph{\strong{units_of_catch}}	Specifies units for catch (2=numbers of fish)

\emph{\strong{se_log_catch}}	Standard error associated with catch

\emph{\strong{Ngenders}}	Number of sexes

\emph{\strong{Nages}}	Number of ages

\emph{\strong{lbin_vector_pop}}	Vector of length bins in the population (cm)

\emph{\strong{lbin_vector}}	Vector of length bins in the data (cm)

\emph{\strong{N_tag_groups}}	Number of tag release events

\emph{\strong{N_recap_events}}	Number of recapture events

\emph{\strong{mixing_latency_period}}	Length of time units (psudo-years) necessary to allow for tag
mixing
}

\section{The lists in \strong{biol_dat} include:}{
\emph{\strong{M:}} age varying natural mortality

\emph{\strong{age}}: first age to last age (pseudo-years)

\emph{\strong{maturity}}: maturity at age ogive (pseudo-years)

\emph{\strong{k}}: age varying growth coefficients

\emph{\strong{L}}: length (cm)

\emph{\strong{Linf}}: length infinity (cm)

\emph{\strong{Lmin}}: length minimum (cm)

\emph{\strong{a}}: length-weight scaling coefficient

\emph{\strong{b}}: length-weight shape parameter

\emph{\strong{LenAge_cv}}: length-age coefficient of variability
}

\examples{
data('YFT_SRD_4A_4_v2')
head(dat_4A_4$catch)
head(biol_dat)

}
