% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YFT_SRD_1A_4}
\alias{YFT_SRD_1A_4}
\title{1 area single representative dataset}
\format{
A data frame 'YFT_SRD_1A_4.RData' with variables:
\describe{
\item{dat_4A_44}{Simulated aggregated 1 area datasets for YFT; Simulation #4}
\item{biol_dat}{Biological data for YFT}
}
}
\usage{
data('YFT_SRD_1A_4')
}
\description{
A single representative datasets aggregated at the 1 area scale. The chosen dataset is simulation #4.
List objects include aggregated catch, cpue, length frequencies, and tagging data (along with biological data).
}
\details{
The key lists in YFT_SRD_1A_$ (e.g., dat_1A_1$lencomp) include:

lencomp: (list) 1 area dataframe of aggregated length frequencies by length bin with columns corresponding to pseudo-year (Yr), season(Seas), fleet/survey index (FltSvy), gender (Gender),  partition – length data come from all fish encountered (Part=0), number of lengths sampled (Nsamp), and the set of length bins (purse seine only)
catch: (list) 1 area dataframe of catch by fishery, including pseudo-year (year), and    season (seas)
CPUE/cpu: (list) 1 area cpue with columns for pseudo-year (year), season (seas), cpue (cpu), standard deviation associated with lognormal error in cpue (cv), and survey index (index); note that CPUE is for longline only
tag_releases: 1 area tag release data with columns corresponding to generic tag release event number (tg), area of release (reg), pseudo-year (yr), season (season), gender (gender), age (age), and number of releases (nrel) adjusted for initial tag mortality
tag_recaps: 1 area tag recapture data with columns corresponding to generic tag release event number (tg), pseudo-year (yr), season (season), fleet index (fleet), and the number of recaptures (recaps)

The lists in biol_dat include:

M: age varying natural mortality
age: first age to last age (pseudo-years)
maturity: maturity at age ogive (pseudo-years)
k: age varying growth coefficients
L: length (cm)
Linf: length infinity (cm)
Lmin: length minimum (cm)
a: length-weight scaling coefficient
b: length-weight shape parameter
}
\examples{
data('YFT_SRD_1A_4')
head(biol_dat)
}
