% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YFT_SRD_4A_4}
\alias{YFT_SRD_4A_4}
\title{4 area single representative dataset}
\format{
A data frame 'YFT_SRD_1A_4.RData' with variables:
\describe{
\item{dat_4A_4}{Simulated aggregated 1 area datasets for YFT; Simulation #4}
\item{biol_dat}{Biological data for YFT}
}
}
\usage{
data('YFT_SRD_4A_4')
}
\description{
A single representative datasets aggregated at the 4 area scale. The chosen dataset is simulation #4.
List objects include aggregated catch, cpue, length frequencies, and tagging data (along with biological data).
}
\details{
Each simulation specific list object contains the following:
}
\examples{
data('YFT_SRD_4A_4')
head(biol_dat)
}
